move eth2 code to its own file to simplify rebase
